+++ App/Page/HomePage.js
@@ -1,53 +1,106 @@
-import React, { Component } from "react";
-import { View, Text, FlatList, ActivityIndicator, Dimensions, TouchableHighlight, StyleSheet } from "react-native";
-import { List, ListItem, SearchBar } from "react-native-elements";
-import { StackNavigator } from 'react-navigation';
-
-var height = Dimensions.get('window').height;
-
-export default class HomePage extends React.Component {
-
-  static navigationOptions = {
-    title: 'Equations',
-  };
-
+import React, { Component } from 'react';
+ 
+import { Text, StyleSheet, View, ListView, TextInput, ActivityIndicator, Alert } from 'react-native';
+ 
+export default class HomePage extends Component {
+ 
   constructor(props) {
+ 
     super(props);
 
     var customData = require('../Model/Equations.json')
     this.equations = customData["Equations"]
+ 
+    this.state = {
+ 
+      isLoading: true,
+      text: '',
+    
+    }
+ 
+    this.arrayholder = [] ;
   }
-
-  renderSeparator = () => {
+ 
+  componentDidMount() {
+         
+        let ds = new ListView.DataSource({rowHasChanged: (r1, r2) => r1 !== r2});
+        this.setState({
+          isLoading: false,
+          dataSource: ds.cloneWithRows(this.equations),
+        }, function() {
+          this.arrayholder = this.equations ;
+         });
+  }
+ 
+  GetListViewItem (name) {
+    
+   Alert.alert(name);
+  
+  }
+  
+   SearchFilterFunction(text){
+     
+     const newData = this.arrayholder.filter(function(item){
+         const itemData = item.name.toUpperCase()
+         const textData = text.toUpperCase()
+         return itemData.indexOf(textData) > -1
+     })
+     this.setState({
+         dataSource: this.state.dataSource.cloneWithRows(newData),
+         text: text
+     })
+ }
+ 
+  ListViewItemSeparator = () => {
     return (
-      <View style={styles.seperator} />
+      <View
+        style={{
+          height: .5,
+          width: "100%",
+          backgroundColor: "#000",
+        }}
+      />
     );
-  };
-
+  }
+ 
+ 
   render() {
-
+    if (this.state.isLoading) {
+      return (
+        <View style={{flex: 1, paddingTop: 20}}>
+          <ActivityIndicator />
+        </View>
+      );
+    }
+ 
     return (
-      <View>
-        <List containerStyle={styles.ListContainer}>
-          <FlatList
-            data={this.equations}
-            renderItem={({ item }) => (
-              // <TouchableHighlight
-              //   style = {{color: 'red'}}>
-                <ListItem
-                  title = {item.name}
-                  titleStyle = {styles.listTitle}
-                  subtitle = {item.equation}
-                  subtitleStyle = {styles.listSubTitle}
-                  containerStyle={styles.lsitItemContainer}
-                  onPress={()=> { this.onPress(item) }}
-                />
-                // </TouchableHighlight>
-              )}
-            keyExtractor={item => item.id}
-            ItemSeparatorComponent={this.renderSeparator}
-          />
-        </List>
+ 
+      <View style={styles.MainContainer}>
+ 
+      <TextInput 
+       style={styles.TextInputStyleClass}
+       onChangeText={(text) => this.SearchFilterFunction(text)}
+       value={this.state.text}
+       underlineColorAndroid='transparent'
+       placeholder="Search Here"
+        />
+ 
+        <ListView
+ 
+          dataSource={this.state.dataSource}
+ 
+          renderSeparator= {this.ListViewItemSeparator}
+ 
+          renderRow={(rowData) => <Text style={styles.rowViewContainer} 
+ 
+          onPress={()=> { this.onPress(rowData) }} > {rowData.name} </Text>}
+ 
+          enableEmptySections={true}
+ 
+          style={{marginTop: 10}}
+ 
+        />
+ 
       </View>
     );
   }
@@ -56,34 +109,31 @@ export default class HomePage extends React.Component {
     this.props.navigation.navigate('Detail', {title: equation.name, equation});
   }
 }
-
+ 
 const styles = StyleSheet.create({
-  seperator:
-    {
-      height: 1,
-      width: "100%",
-      backgroundColor: "#CED0CE",
-      marginLeft: "5%"
-    },
-  ListContainer:
-  {
-    height,
-    marginTop: 0,
-    borderTopWidth: 0,
-    borderBottomWidth: 0
+ 
+ MainContainer :{
+ 
+  justifyContent: 'center',
+  flex:1,
+  margin: 7,
+ 
   },
-  lsitItemContainer:
-  {
-    borderBottomWidth: 0
+ 
+ rowViewContainer: {
+   fontSize: 17,
+   padding: 10
   },
-  listTitle:
-  {
-    fontSize: 30,
-    paddingTop: 10,
-    paddingBottom: 10
-  },
-  listSubTitle:
-  {
-    fontSize: 24
-  },
+ 
+  TextInputStyleClass:{
+        
+   textAlign: 'center',
+   height: 40,
+   borderWidth: 1,
+   borderColor: '#009688',
+   borderRadius: 7 ,
+   backgroundColor : "#FFFFFF"
+        
+   }
+ 
 });
